excel_links_opener/requirements.txt
# Required packages for Excel Links Opener

pandas>=1.3.0
openpyxl>=3.0.0
pyyaml>=6.0
